Continuing the integer datatype conversion.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@251 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/s_slit.c
@@ -492,7 +492,7 @@ mxs_slit_scan_motor_record_array_move_special(
 			default:
 				return mx_error( MXE_CORRUPT_DATA_STRUCTURE,
 					fname, "Unknown slit type %ld.",
-					slit_motor->slit_type );
+					(long) slit_motor->slit_type );
 				break;
 			}
